Implementing shared source code from dependencies
project.json may contain "shared": "{pattern}"

The default value for this project field is "Compiler\Shared\**\*.cs"

Files matching this pattern are not used for compilaion, instead they
are added to the compilation of projects with a direct dependency.

These files are also put into the nupkg in a "shared\*" path. A direct
dependency on a nupkg with shared\* source will also cause the source
to be compiled in.