Use mono or mono symlink if present
This will help in cases where KRE is installed on a machine to use a
custom build of mono that doesn’t interfere with the standard mono
release. That works by going into the KRE folder and creating a symlink
to exactly the mono executable that should be used to host k commands.