Enable to script block in Mono environment
Rules:

In non-mono environment, if the ComSpec exists the script block will
wrapped in cmd /C, otherwise the first argument will be used as FileName
for the new process.

For mono environment if the given script path can be found after appending
.sh to it then execute that script with /bin/bash. Otherwise use the
script name as FileName for process start info.