Fixing preprocess code failing to load a class library
This fixes bug : https://github.com/aspnet/XRE/issues/1048

On Mono while loading the !preprocess assembly the ResolveEventArgs.RequestingAssembly is not populated.
As a result when code in !preprocess references a class library the load fails on the class library.
Below is a work around to try populate the requesting assembly.