Remove dead code
The "Firmware" methods living in module.cpp for the coreclr native
loader appeared to be unused by the klr bootstrapper. I have removed
them and updated the call in klr.cpp to not use the firmware name.

I was able to use klr after building with this change to run hello
world against coreclr, so I don't believe this stuff is needed
anymore.