Fix size_t to int conversion warnings
Some code would convert size_t (which is unsigned) to int.  This could
lead to overflow.  In practice, we are unlikely to hit these errors
since all the input strings are bounded by MAX_PATH, but we might as
well just write the code in a way such that it works correctly.