Use LPCTSTR in tpa.cpp
In order to share tpa.cpp between Unix and Windows, we start using
PTCTSTR in favor of wide strings everywhere.  While I am touching the
file, I am cleaning up the use of _wcstrdup.  The major issue with this
code was the use of delete [] to free the memory, which is incorrect
since _wcstrdup allocated memory using malloc.

There's no reason for these strings to not be const, so we can
just use LPCTSTR and hand back data backed by string literals.