Refactor string comparision helpers
There were a few string comparision helpers in the code that passed
MAX_PATH to strnicmp.  Because of this, these functions were not general
purpose functions, but they were named that way, which folks (myself
included) found confusing.

This change uses the term "Path" in functions which are working over
paths.  The more general purpose StringsEqual assumes null terminated
strings.