Fix dnx tracing message
In f6e7bddc I converted dnx to use _t methods and macros to switch
between wide and ansi strings depending on if we were building for
Windows or Unix.

When I did the pass over dnx, I incorrectly converted a printf_s to
_tprintf_s even though it always uses narrow strings (by design).  This
causes the printf call to print garbage data for the string argument,
since it tries to treat it on Windows as a wide string.

Simply revert this one line back to printf_s.