Load libcoreclrpal before loading libcoreclr
Recent changes to libcoreclr require loading the pal before loading the
runtime itself.

The runtime will probabbly move away from this model in a few months, so
the code is structured in a way such that if we can't load the PAL we
still try to load the runtime.

Fixes #1663