Returning error codes instead of exceptions (Fixes #1825)
Exceptions thrown in managed code cannot be handled in native code and are causing hard crashes. So, if no one handled an exception we will catch it and convert into an error code. We do it in methods that are invoked from native code. Note that at the moment we will do it only for Core Clr. On full Clr the exceptions are handled and in Helios we already have Clr up and running and are able to catch exceptions.